closes #733 refs #657
The scatter selection can work based on both point (click) or rectangle (drag the mouse), however the logic used to acquire the selected point needs to be refined, the qwt's closestPoint function does not perform well enough.

A signal/slot mechanism is in use so that the table can notify the application that a chart is being created.