closes #793
The wizard was deemed unnecessary so a new dialog was created to replace it. The proccess is now much simpler than with the wizard.

Changes were made to the Query Laye, SQL visitor and serializer in order to ensure their functions would work with Joins. Further testing is required.

OPtimmizations to the dialog are under developmente, a new ticket related to these enhancements will be oppened.