closes #823
Adding a new function to the dataaccess module that returns the name of the base dataset from a linked dataset.
Adding a new attribute to the property class that stores the name of it's original dataset.
Several changes to the histogram related functions, the summary functions have been included and required various changes to it's creation functions and widgets.
Added a temporary icon to the query layer to act as a placeholder until a better icon can be found.