closes #867
A new checkbox has been added to the scatter data widget in order to allow the user to choose weather or not to use the whole data from a raster while creating a scatter, if not, 1000 pixels will be selected at random and the scatter will be based on those pixels.

Therefore, it enables the user to choose to prioritize either speed or accuracy.