closes #864, refs #867
Minor change to the scatter limit, if the user does not want to wait for the whole scatter to be processed, instead of using a hard-coded value the maximum number of points is calculated on runtime.

A new tooltip and a new note on the help files have been added to highlight the fact that, when a histogram is based on a raster, setting 0 slices will proccess all of the pixel data.
The default number of slices for raster based histogram has been changed to 30, as used by other appications, in order to prevent processing the whole data by default.