closes #1037
The function te::da::DataSourceInfoManager::add was verifying only if the id of a given dataSource was unique, since every new dataSource - including those with identical connection infos -  contained a randomly generated id, they all passed the check.

A new verification was included, it checks if the connection info is identical. Minor changes were made to the connectors to check weather or not a new datasource was successfully included.