refs #1037
Further enhancements related to this bugfix. If the user tries to add a datasource with the same connection info of a previously defined datasource, then the datasource that is already defined will be used instead.
This was done to prevent layers from being created with a reference to a dataSource that was never added to the dataSourceInfoManager.