refs #1066
The table link dialog was not using the scopedCursor, this has been changed in order to prevent errors related to the regular cursor not being restored.

refs #1069

A tooltip has been created to layers of the type QueryLayer.
The queryLayer has a new function that returns it's restriction as a string.